This is  a function that frees a listint_t list.
Prototype: void free_listint(listint_t *head); - The function takes in one argument: a pointer to the head of the list (head).

A temporary listint_t pointer variable, temp, is created and initialized to the value of the head pointer.

The function then enters a while loop, which continues as long as the temp pointer is not NULL. Inside the loop, the head pointer is updated to point to the next node in the list, and the temp pointer is freed using the free() function. This frees the memory that was allocated for the node.

Finally, the function returns nothing, since it's a void function.

By repeatedly updating the head pointer and freeing each node in turn, this function effectively frees the entire linked list from memory.

It's important to note that this function assumes that the list was created using dynamic memory allocation, and that each node was created using malloc() or a similar memory allocation function. If the list was created using automatic memory allocation (i.e. on the stack), or if some other type of memory allocation function was used, this function may not work correctly.