This is a function that returns the sum of all the data (n) of a listint_t linked list.
Prototype: int sum_listint(listint_t *head);
if the list is empty, return 0
Details
The function takes in one argument: a pointer to the head of the linked list (head).

A variable called sum_data is created and initialized to 0.

The function then iterates through the linked list using a while loop. For each node in the list, it adds the value of the node's data (n) to the sum variable, and updates the head pointer to point to the next node in the list.

Once the loop has completed, the function returns the value of the sum_data variable, which represents the total sum of all the data (n) in the linked list.

It's important to note that this function assumes that the list was created using dynamic memory allocation, and that each node was created using malloc() or a similar memory allocation function. If the list was created using automatic memory allocation (i.e. on the stack), or if some other type of memory allocation function was used, this function may not work correctly.