This is a function that inserts a new node at a given position.
Prototype: listint_t *insert_nodeint_at_index(listint_t **head, unsigned int idx, int n);
where idx is the index of the list where the new node should be added. Index starts at 0
Returns: the address of the new node, or NULL if it failed
if it is not possible to add the new node at index idx, do not add the new node and return NULL
Details:
This code  implement a function called insert_nodeint_at_index() that inserts a new node containing a given integer n at a specified index idx in a singly linked list.

The function takes a double pointer head which points to the pointer to the head of the linked list. It returns a pointer to the newly created node or NULL if the operation failed.

The function first traverses the linked list to reach the node at index (idx-1) or at the beginning of the list if idx is zero. If the traversal fails, it returns NULL.

It then creates a new node, sets its n field to the input n, and inserts it into the linked list by updating the next pointers of the adjacent nodes.

The function then returns a pointer to the newly created node.