This is a function that deletes the node at index index of a listint_t linked list.
Prototype: int delete_nodeint_at_index(listint_t **head, unsigned int index);
where index is the index of the node that should be deleted. Index starts at 0
Returns: 1 if it succeeded, -1 if it failed
Details.
This program defines a function delete_nodeint_at_index that deletes the node at a given index in a linked list.

The function takes in the address of the head of the linked list and the index of the node to be deleted. It returns 1 if the deletion succeeded and -1 if it failed.

The function works by traversing the linked list to find the node at the specified index. Once found, it adjusts the previous node's next pointer to point to the next node, effectively removing the node at the specified index. If the node to be deleted is the head of the linked list, the function updates the head pointer to point to the next node instead. Finally, the function frees the memory allocated for the deleted node using the free() function.

This program assumes that the linked list is indexed starting from 0, and that the indices are sequential with no gaps. If the index provided is greater than or equal to the number of nodes in the linked list, or if the linked list is empty, the function returns -1 to indicate failure