This is a function that frees a listint_t list.
Prototype: size_t free_listint_safe(listint_t **h);
This function can free lists with a loop
You should go though the list only once
Returns: the size of the list that was free’d
The function sets the head to NULL
Details:
These functions implement a modified version of a linked list data structure in C. The listint_t structure represents a node of a linked list, where each node contains an integer value and a pointer to the next node in the list. The listp_t structure is used to create a linked list of pointers to listint_t nodes in order to keep track of which nodes have already been printed or freed.

The function print_listint_safe prints the values of the nodes in the linked list pointed to by head, while also checking for loops in the list. If a loop is found, it prints an arrow pointing to the node that starts the loop and returns the number of nodes that were printed before the loop was detected. If no loop is found, it prints all of the nodes in the list and returns the total number of nodes printed.

The function free_listint_safe frees the nodes in the linked list pointed to by h, while also checking for loops in the list. If a loop is found, it frees all of the nodes up to the start of the loop (including the start node) and returns the number of nodes that were freed before the loop was detected. If no loop is found, it frees all of the nodes in the list and returns the total number of nodes freed.

Both functions use the listp_t structure to keep track of which nodes have already been processed. When a node is printed or freed, its pointer value is stored in a listp_t node and added to the linked list pointed to by hptr. Before printing or freeing a node, the listp_t list is checked for the presence of the node's pointer value. If the pointer is found in the list, it indicates that the node is part of a loop and the appropriate action is taken. Otherwise, the node is printed or freed and its pointer value is added to the list.