Test code for 0-binary_to_uint.c.
This code is testing the binary_to_uint function. It converts a binary string to an unsigned integer.

The program calls the function several times, passing different binary strings as arguments, and then prints the results using the printf function.

The expected output is a series of unsigned integers that correspond to the binary strings that were passed as arguments.