a function that converts a binary number to an unsigned int.
Prototype: unsigned int binary_to_uint(const char *b);
where b is pointing to a string of 0 and 1 chars
Return: the converted number, or 0 if
there is one or more chars in the string b that is not 0 or 1
b is NULL
This program takes in a binary string and converts it to an unsigned integer.

The function binary_to_ant starts by initializing variables a, len, and base_two. a will hold the unsigned integer eqavalent of the binary string, len will hold the length of the string, and base_two will be used to calculate the binary value of each digit.

The program then checks if the input is NULL and returns 0 if true.

The program then iterates through the string b from the last character to the first character, updating the value of base_two with each iteration. If the character is not a 0 or 1, the function returns 0. If the character is a 1, the program updates the value of a with the value of base_two.

The function then returns a.