a function that prints the binary representation of a number.
Prototype: void print_binary(unsigned long int n);
Format: see example
You are not allowed to use arrays
You are not allowed to use malloc
You are not allowed to use the % or / operators
This implementation of print_binary uses recursion to print the binary representation of the input number.

The function first checks if the input number is greater than 1. If it is, it recursively calls itself with the input number right-shifted by 1 (effectively dividing by 2). This allows the function to print the binary representation from the most significant bit to the least significant bit.

Once the base case is reached (when the input number is less than or equal to 1), the function uses the bitwise AND operator (&) with 1 to determine the value of the least significant bit. It then adds the character value of '0' to that bit (either 0 or 1) to convert it to a printable character and writes it to the standard output using _putchar.

By using recursion, this implementation does not require the use of arrays or loops, and it avoids the use of the division and modulus operators (/ and %) that would normally be used to convert the number to binary.