a function that returns the value of a bit at a given index.
Prototype: int get_bit(unsigned long int n, unsigned int index);
where index is the index, starting from 0 of the bit you want to get
Returns: the value of the bit at index index or -1 if an error occured
This implementation of get_bit looks correct. It checks if the input n is 0 and the index is less than 64, and returns 0 if it is true. This is because if n is 0, all the bits in it are 0, and accessing any bit would result in undefined behavior. The loop then iterates through all the bits in n, shifting it to the right by 1 bit in each iteration, and checking if the current index matches the input index. If it does, the function returns the value of the bit at that index. If it reaches the end of the loop without finding the index, it returns -1 to indicate an error.