a function that sets the value of a bit to 1 at a given index.
Prototype: int set_bit(unsigned long int *n, unsigned int index);
where index is the index, starting from 0 of the bit you want to set
Returns: 1 if it worked, or -1 if an error occurred
It takes a pointer to an unsigned long int n and an index index, sets the bit at index index to 1, and returns 1 if it worked, or -1 if an error occurred. It first checks if the index is greater than 63, in which case it returns -1 since the unsigned long int has 64 bits. Otherwise, it creates a mask by left-shifting 1 by the index to set the bit to 1. Finally, it performs a bitwise OR operation between the mask and the value pointed to by n to set the bit at the given index to 1.