a function that sets the value of a bit to 0 at a given index.
Prototype: int clear_bit(unsigned long int *n, unsigned int index);
where index is the index, starting from 0 of the bit you want to set
Returns: 1 if it worked, or -1 if an error occurred
The function first checks if the given index is greater than 63, which is the maximum index for a 64-bit unsigned long int. If the index is out of range, the function returns -1 to indicate an error.

If the index is within range, the function creates a bit mask with a 1 at the given index using the left shift operator <<. This sets the k-th bit to 1 and all other bits to 0. The variable k is used to store this bit mask.

Next, the function checks if the k-th bit of the given number *n is already set to 1 using the bitwise AND operator &. If the k-th bit is already set to 1, the function clears the bit at the given index by using the bitwise XOR operator ^ with the bit mask k. The result of this operation is that the k-th bit of *n is set to 0 while all other bits remain unchanged.

Finally, the function returns 1 to indicate success.

Note that this implementation uses the XOR operator instead of the AND operator with the bitwise complement to clear the bit. This is because the XOR operator is a more efficient way to toggle a bit in C, as it does not require a separate complement operation.