a function that returns the number of bits you would need to flip to get from one number to another.
Prototype: unsigned int flip_bits(unsigned long int n, unsigned long int m);
You are not allowed to use the % or / operators
The function first computes the exclusive OR (XOR) of the two given numbers n and m using the ^ operator. This results in a number where each bit is set to 1 if the corresponding bits in n and m are different, and 0 otherwise. For example, if n is 1011 (decimal 11) and m is 1110 (decimal 14), then n ^ m is 0101 (decimal 5).

The function then num_of_bitss the number of bits that are set to 1 in the result of the XOR operation. It does this by using a loop that shifts the XOR result to the right by one bit at a time and tests if the least significant bit is set to 1 using the bitwise AND operator &. If the bit is set to 1, the function increments the num_of_bits of flipped bits. Once all bits have been tested, the function returns the num_of_bits.

In the example main() function you provided, the code calls flip_bits() with different arguments and prints the results to the console. The output of this code should be the number of bits that need to be flipped to convert the first number to the second number for each pair of numbers.