a function that checks the endianness.
Prototype: int get_endianness(void);
Returns: 0 if big endian, 1 if little endian.
This function checks the endianness of the machine by creating an unsigned int variable r with the value 1, and then casting a pointer to r  to a char pointer c. The char pointer is used to access the individual bytes of the unsigned int value in memory.

If the first byte of the unsigned int value is non-zero, then the machine is little-endian, because the least significant byte is stored first in memory. If the first byte is zero, then the machine is big-endian, because the most significant byte is stored first in memory.