Add test code for 0-read_textfile() function
An implementation of the main() function that calls the read_textfile() function with different arguments and prints the result to the console.

The code expects the program to be executed with a command-line argument, which should be the name of a file. If the number of command-line arguments (ac) is not equal to 2, the code displays an error message using dprintf() and exits with a non-zero status code.

If the number of command-line arguments is 2, the code calls read_textfile() with the file name (av[1]) and a specified number of letters to read (114 and 1024 in this case). The returned value from read_textfile() is assigned to the variable n.

After calling read_textfile(), the code prints a message indicating the number of characters printed by read_textfile().

Note that the implementation of the read_textfile() function is not provided in the code snippet you provided. In order to understand the behavior and expected output of the program, you would need to examine the implementation of the read_textfile() function.